backend/requirements.txt
attrs==19.1.0
blis==0.2.4
certifi==2019.6.16
chardet==3.0.4
cymem==2.0.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.2
django-allauth==0.39.1
django-cors-headers==3.0.2
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.9.4
gunicorn==19.9.0
idna==2.8
jsonschema==3.0.1
murmurhash==1.0.2
nltk==3.4.3
numpy==1.16.4
oauthlib==3.0.1
plac==0.9.6
preshed==2.0.1
psycopg2==2.8.3
pyrsistent==0.15.2
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
spacy==2.1.4
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.1.0/en_core_web_sm-2.1.0.tar.gz#egg=en_core_web_sm==2.1.0
sqlparse==0.3.0
srsly==0.0.7
thinc==7.0.4
tqdm==4.32.2
urllib3==1.25.3
wasabi==0.2.2
whitenoise==4.1.2
